requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
tavily-python==0.5.0
supabase==2.28.3
python-multipart==0.0.12
Pillow==10.4.0
